Normalized and balanced Poisoned Blades skill. Improved Fixative skill.
Poisoned Blades are adjusted in a way that the skill depends much less on
oil level. The poison apply chance is 6% per skill level and gets +5%,+10%
apply chance increase with oils level 2,3. Also the power of the poison
from this skill is reduced by about 37%. The outcome is that the skill is
stronger in early game (because of higher applu chance), but a bit weaker
finally.

Fixative skill now also gives a chance to apply a random effect on oil
vulnerable enemy. The apply chance is 12% per skill level. The possible
effects are currently: burning, bleeding, blindness, poison (stronger than
from Poisoned Blades).